Moved app.tsx + Added confirmation message if you leave the page without saving the config
